Fix cleanChallenges running with an error

+++ src/repositories/ChallengeRepository.ts
@@ -9,6 +9,7 @@ import { LoggerService } from "../services/LoggerService";
 export class ChallengeRepository extends Repository<Challenge>  {
     
     log = Container.get(LoggerService);
+    expireTime = Number(process.env.SESSION_EXPIRY_IN_MINUTES) * 60000;
 
     public async createChallenge(user: User): Promise<string> {
         const challenge = new Challenge();
@@ -30,17 +31,21 @@ export class ChallengeRepository extends Repository<Challenge>  {
     }
 
     public async cleanChallenges(): Promise<void> {
-        const expireTime = Number(process.env.SESSION_EXPIRY_IN_MINUTES) * 60000;
-
-        const repo = await getRepository(Challenge);
-        const challenges = await repo.find({
-            where: {
-                "createdAt": Raw(alias => `${Date.now().toString()} - ${alias} > ${expireTime}`)
+        try {
+            const repo = await getRepository(Challenge);
+            const challenges = await repo.find({
+                where: {
+                    "createdAt": Raw(alias => `${Date.now().toString()} - ${alias} > ${this.expireTime}`)
+                }
+            });
+            if (challenges.length > 0) {
+                this.log.info(`Deleting ${challenges.length} stale challenges...`);
+                for (const challenge of challenges) {
+                    repo.delete(challenge);
+                }
             }
-        });
-        console.log(`Deleting ${challenges.length} stale challenges...`);
-        for (const challenge of challenges) {
-            repo.delete(challenge);
+        } catch (e) {
+            this.log.error(e);
         }
     }
